+++ src/anchored-formula.jl
@@ -296,42 +296,3 @@ function parseformula(
 )
     AnchoredFormula(logic, parseformula(SyntaxTree, expr, operators(logic); kwargs...))
 end
-
-"""$(doc_randformula)"""
-function randbaseformula(
-    height::Integer,
-    g::AbstractGrammar;
-    kwargs...
-)::AnchoredFormula
-    _alphabet = alphabet(g)
-    _operators = operators(g)
-    baseformula(
-        randformula(height, _alphabet, _operators; kwargs...);
-        alphabet = _alphabet,
-        additional_operators = _operators
-    )
-end
-
-function randbaseformula(
-    height::Integer,
-    alphabet,
-    operators::AbstractVector{<:Operator};
-    kwargs...
-)::AnchoredFormula
-    alphabet = convert(AbstractAlphabet, alphabet)
-    baseformula(
-        randformula(height, alphabet, operators; kwargs...);
-        alphabet = alphabet,
-        additional_operators = operators,
-    )
-end
-
-function randbaseformula(
-    height::Integer,
-    g::AbstractGrammar,
-    args...;
-    rng::Union{Integer,AbstractRNG} = Random.GLOBAL_RNG,
-    kwargs...
-)::AnchoredFormula
-    randbaseformula(height, alphabet(g), operators(g), args...; rng=rng, kwargs...)
-end
