+++ src/generation/formula.jl
@@ -210,6 +210,7 @@ end
     modaldepth::Integer=height,
     atompicker::Union{Function,AbstractWeights,AbstractVector{<:Real},Nothing}=randatom,
     opweights::Union{AbstractWeights,AbstractVector{<:Real},Nothing}=nothing,
+
     kwargs...
 )
     rng = initrng(rng)
@@ -304,4 +305,44 @@ end
     kwargs...
 )
     randformula(initrng(rng), height, args...; kwargs...)
+end
+
+
+function randbaseformula(
+    height::Integer,
+    g::AbstractGrammar;
+    kwargs...
+)::AnchoredFormula
+    _alphabet = alphabet(g)
+    _operators = operators(g)
+    baseformula(
+        randformula(height, _alphabet, _operators; kwargs...);
+        alphabet = _alphabet,
+        additional_operators = _operators
+    )
+end
+
+""""""
+function randbaseformula(
+    height::Integer,
+    alphabet,
+    operators::AbstractVector{<:Operator};
+    kwargs...
+)::AnchoredFormula
+    alphabet = convert(AbstractAlphabet, alphabet)
+    baseformula(
+        randformula(height, alphabet, operators; kwargs...);
+        alphabet = alphabet,
+        additional_operators = operators,
+    )
+end
+
+function randbaseformula(
+    height::Integer,
+    g::AbstractGrammar,
+    args...;
+    rng::Union{Integer,AbstractRNG} = Random.GLOBAL_RNG,
+    kwargs...
+)::AnchoredFormula
+    randbaseformula(height, alphabet(g), operators(g), args...; rng=rng, kwargs...)
 end
